+++ backend/src/main/java/com/scanpilot/auth/config/AuthConfigProperties.java
@@ -1,8 +1,12 @@
 package com.scanpilot.auth.config;
 
+import lombok.Getter;
+import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 
+@Getter
+@Setter
 @Configuration
 @ConfigurationProperties(prefix = "scanpilot.auth")
 public class AuthConfigProperties {
@@ -17,68 +21,4 @@ public class AuthConfigProperties {
     private long sessionTtlSeconds = 604800; // 7 days
     private long stateTtlSeconds = 600; // 10 minutes
     private String cookieName = DEFAULT_SESSION_COOKIE_NAME;
-
-    public String getClientId() {
-        return clientId;
-    }
-
-    public void setClientId(String clientId) {
-        this.clientId = clientId;
-    }
-
-    public String getClientSecret() {
-        return clientSecret;
-    }
-
-    public void setClientSecret(String clientSecret) {
-        this.clientSecret = clientSecret;
-    }
-
-    public String getRedirectUri() {
-        return redirectUri;
-    }
-
-    public void setRedirectUri(String redirectUri) {
-        this.redirectUri = redirectUri;
-    }
-
-    public String getFrontendUrl() {
-        return frontendUrl;
-    }
-
-    public void setFrontendUrl(String frontendUrl) {
-        this.frontendUrl = frontendUrl;
-    }
-
-    public boolean isCookieSecure() {
-        return cookieSecure;
-    }
-
-    public void setCookieSecure(boolean cookieSecure) {
-        this.cookieSecure = cookieSecure;
-    }
-
-    public long getSessionTtlSeconds() {
-        return sessionTtlSeconds;
-    }
-
-    public void setSessionTtlSeconds(long sessionTtlSeconds) {
-        this.sessionTtlSeconds = sessionTtlSeconds;
-    }
-
-    public long getStateTtlSeconds() {
-        return stateTtlSeconds;
-    }
-
-    public void setStateTtlSeconds(long stateTtlSeconds) {
-        this.stateTtlSeconds = stateTtlSeconds;
-    }
-
-    public String getCookieName() {
-        return cookieName;
-    }
-
-    public void setCookieName(String cookieName) {
-        this.cookieName = cookieName;
-    }
 }